[LEMS-4071/too-static] move getEquationString to its own file
